+++ apps/pong/CHANGELOG.md
@@ -1,5 +1,15 @@
 # inglorious-pong
 
+## 1.0.4
+
+### Patch Changes
+
+- Support eager and lazy loading of images
+- Updated dependencies
+  - @inglorious/engine@2.2.0
+  - @inglorious/renderer-2d@1.0.0
+  - @inglorious/utils@3.5.3
+
 ## 1.0.3
 
 ### Patch Changes