+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/eslint-config
 
+## 1.0.1
+
+### Patch Changes
+
+- Support eager and lazy loading of images
+
 ## 1.0.0
 
 ### Major Changes